Generate reproducible CLIs via go-subcommand
- Isolates application-specific CLI and Shiny GUI imports into `internal/cli`,
  preserving the public `md2png` package as a lightweight library.
- Migrates workflow documentation and testing entirely to GNU double-dash
  long-flag syntax to conform cleanly with the `go-subcommand` generator.
- Integrates reproducible `gosubc@v0.0.29` code-generation, cleanly omitting
  buggy `root_test.go` stubs automatically while maintaining deterministic builds.
- Refactors tests against the final pointer-driven conversion seam.
