+++ AGENTS.md
@@ -1,3 +1,3 @@
 # Agent Instructions
 
-After modifying the project, run `go run . -in README.md -out output.png` to regenerate the sample output, and inspect `output.png` to verify the rendering looks correct.
+After modifying the project, run `go run . --in README.md --out output.png` to regenerate the sample output, and inspect `output.png` to verify the rendering looks correct.
