+++ cmd/md2view/main.go
@@ -20,7 +20,7 @@ var (
 // Provenance
 var (
 	GeneratorVersion = "v0.0.29"
-	GeneratedAt      = "2026-09-08T04:57:06Z"
+	GeneratedAt      = "2026-09-08T08:22:07Z"
 	ProjectCommit    = "1bbcbac0f44ab4a63680557f16ec90a825bbf6cb"
 )
