+++ generate.go
@@ -1,4 +1,4 @@
 package md2png
 
-//go:generate sh -c "command -v gosubc >/dev/null 2>&1 && gosubc generate || go run github.com/arran4/go-subcommand/cmd/gosubc@latest generate"
-//go:generate sh -c "rm -f cmd/md2png/root_test.go cmd/md2view/root_test.go"
+//go:generate go run github.com/arran4/go-subcommand/cmd/gosubc@v0.0.29 generate
+//go:generate rm -f cmd/md2png/root_test.go cmd/md2view/root_test.go
